Projekt2/requirements.txt
mpmath==1.3.0
numpy==1.26.4
pillow==10.3.0
pycryptodome==3.20.0
sympy==1.12.1
